camera ready

+++ testing.py
@@ -8,19 +8,15 @@ import os
 import pandas as pd 
 from data.dataset import GetDataset
 
-loss = ["WS",] #["QR","WS","RQR-W","SQR","RQR-O","OQR","IR"]
 
+loss = ["RQR-W","SQR","SQRN","IR","WS"]
 
 def testing():
     parser = argparse.ArgumentParser(description='Run the experiment')
     parser.add_argument('--dataset_name', type=str, default="boston", help='The name of the dataset')
     parser.add_argument('--n_seed', type=int, default=10, help='Number of seed')
-    parser.add_argument('--loss', type=str, default="QR", help='The name ofthe loss function')
     parser.add_argument('--gpu', type=int, default=0, help='The id of the gpu to use')
     parser.add_argument('--coverage', type=float, default=0.9, help='Targeted coverage')
-
-    output_path = "results/finetuning/test_results_simul.csv"
-
 
     args = parser.parse_args()
     
@@ -38,15 +34,13 @@ def testing():
     
     data = GetDataset(config["dataset_name"], "data")
 
-    l = args.loss
-
-    config["loss"] = l
-    if l == "SQRN":
-        l = "SQR"
 
     
 
     for l in loss:
+        
+        output_path = f"results/finetuning/results_{args.dataset_name}_{l}.csv"
+
         config["loss"] = l
         if l == "SQRN":
             l = "SQR"
@@ -57,13 +51,10 @@ def testing():
             penalty = fine_tuned_parameters["penalty"]    
         
         for seed in range(args.n_seed):
-            if not os.path.exists(output_path):
-                pd.DataFrame([result_dict]).to_csv(output_path)
-            else:
-                old_df = pd.read_csv(output_path,index_col=0)
+        
 
-            if old_df[ (old_df["dataset"] == args.dataset_name) & (old_df["loss"] == l) & (old_df["seed"] == seed) & (old_df["target"] ==config["coverage"] )].shape[0] > 0:
-                continue
+            # if old_df[ (old_df["dataset"] == args.dataset_name) & (old_df["loss"] == l) & (old_df["seed"] == seed) & (old_df["target"] ==config["coverage"] )].shape[0] > 0:
+            #     continue
             
             
             config["random_seed"] = seed
@@ -73,15 +64,15 @@ def testing():
             tolerance = 2.5/100
             best_config  = {}
             for p in penalty:
-                study = optuna.load_study(storage= "sqlite:///results/finetuning/recording.db", study_name = args.dataset_name+ "_"+ l + "_" + str(p) + "_" + str(seed))
+                study = optuna.load_study(storage= "sqlite:///results/finetuning/recording.db", study_name = args.dataset_name+ "_" +  str(args.coverage) + "_" + l + "_" + str(p) + "_" + str(seed))
                 best_params = study.best_params
                 for key in best_params.keys():
                     config[key] =  best_params[key]
                 config["penalty"] = p
 
                 val_loss,results, results_val = run_experiment(config, data)
-                # if (results_val["test_coverage"] != study.best_trial.user_attrs["coverage"] or results_val["test_width"] != study.best_trial.user_attrs["width"]) and config["loss"] != "SQRN":
-                #     raise ValueError("The best value is not the same")
+                if (results_val["test_coverage"] != study.best_trial.user_attrs["coverage"] or results_val["test_width"] != study.best_trial.user_attrs["width"]) and config["loss"] != "SQRN":
+                    raise ValueError("The best value is not the same")
                 
                 
                 if results["test_coverage"] < config["coverage"]*(1 + tolerance) and results["test_coverage"] > config["coverage"]*(1 - tolerance):
@@ -114,9 +105,12 @@ def testing():
             for key in best_config.keys():
                 result_dict[key] = best_config[key]
 
-            old_df = pd.read_csv(output_path,index_col=0)
-            df = pd.DataFrame([result_dict])
-            pd.concat([old_df, df]).to_csv(output_path)
+            if not os.path.exists(output_path):
+                pd.DataFrame([result_dict]).to_csv(output_path)
+            else:
+                old_df = pd.read_csv(output_path,index_col=0)
+                df = pd.DataFrame([result_dict])
+                pd.concat([old_df, df]).to_csv(output_path)
         
 if __name__ == "__main__":
     testing()
